+++ doc/code/screens.rb
@@ -7,9 +7,9 @@
 
 # ** Render the screen. **
 #
-# You should rewrite the `render` method of `Prime::BaseScreen`, which will be runned after first opening screen.
+# You should rewrite the `render` method of `Prime::Screen`, which will be runned after first opening screen.
 
-class FooScreen < Prime::BaseScreen
+class FooScreen < Prime::Screen
   def render
     @main_section = MyProfileSection.new(screen: self, model: User.first)
     @main_section.render
@@ -22,13 +22,13 @@ end
 #
 # NOTE: screen should be created with enabled navigation (see "Initialize screen" block).
 
-class FooScreen < Prime::BaseScreen
+class FooScreen < Prime::Screen
   title 'Foo screen'
 end
 
 # You can pass block to define screen's title
 
-class FooScreen < Prime::BaseScreen
+class FooScreen < Prime::Screen
   title { params[:title] }
 end
 
@@ -51,7 +51,7 @@ end
 # You can use root: true when you have already opened screen with sidebar, and you want to open new screen without sidebar.
 # This option is false by default if you already have root screen and true if not.
 #
-# * `:sidebar`. Send `Prime::BaseScreen` instance to this option if you want to create root screen with sidebar. 
+# * `:sidebar`. Send `Prime::Screen` instance to this option if you want to create root screen with sidebar. 
 # Value of this options will be used as sidebar controller.
 
 def open_foo_screen